Fix construction from empty arrays in debug modes
Use .data() instead of &arr[0] to get the array pointer. This avoids
calling operator[] with 0, which causes assertions in some debug modes
to fail. For example when using -D_GLIBCXX_DEBUG for the libstdc++ debug
mode. In Visual Studio, a similar debug mode is even the default for
Debug builds, so this problem can occur quite easily.