Fix remaining cases of -Wtype-limits warning (#62)
Following up on 1a36383 by using detail::is_positive to fix a few
more cases of the warning -Wtype-limits with gcc 10. Curiously,
the warning does not appear for the checks of the template
parameters Count and Offset of the static variants of first, last
and subspan at the moment. But does not hurt to use is_positive for
those too.